2nd pass of shadow mapping is working, except  is still unset, fixing 1st pass

+++ Source/fps_counter.cpp
@@ -19,8 +19,8 @@ void FPS_Counter::get_end_frequency()
 
 double FPS_Counter::get_frametime()
 {
-	double t_start = count_start.QuadPart;
-	double t_end = count_ende.QuadPart;
+	double t_start = (double) count_start.QuadPart;
+	double t_end = (double) count_ende.QuadPart;
 	double ms_per_frame = (t_end - t_start) / (double)frequenz.QuadPart;
 	double fps = 1.0 / (ms_per_frame / 1000.0);
 	return ms_per_frame;
